# artifacts/test_baseline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GesahniV2 API",
    "version": "v0.5.0-41-gd33726d"
  },
  "paths": {
...
    "/v1/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Full Status",
        "operationId": "full_status_v1_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Full Status V1 Status Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2": [
              "admin:write"
            ]
          }
        ]
      }
    },
...
    "/v1/admin/tokens/google": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Tokens Google",
        "description": "Admin view: compact tokens for a given user_id (auth-gated).\n\nReturns a compact list of token rows (valid and invalid) for inspection.",
        "operationId": "admin_tokens_google_v1_admin_tokens_google_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/music": {
      "post": {
        "tags": [
          "Music"
        ],
        "summary": "Music Command",
        "operationId": "music_command_v1_music_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ui/duck": {
      "post": {
        "tags": [
          "TV"
        ],
        "summary": "Ui Duck",
        "operationId": "ui_duck_v1_ui_duck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuckBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DuckBody": {
        "properties": {
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 15
          }
        },
        "type": "object",
        "title": "DuckBody",
        "example": {
          "level": 15
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MusicCommand": {
        "properties": {
          "command": {
            "type": "string",
            "title": "Command",
            "description": "play|pause|next|previous|volume"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "temporary": {
            "type": "boolean",
            "title": "Temporary",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "MusicCommand",
        "example": {
          "command": "volume",
          "temporary": true,
          "volume": 20
        }
      },
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "OkResponse",
        "example": {
          "ok": true,
          "status": "ok"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "care:resident": "Resident-level care features (presence, sessions, HA actions)",
              "care:caregiver": "Caregiver portal and actions",
              "music:control": "Control music playback and devices",
              "admin:write": "Administrative endpoints: flags, metrics, backups"
            },
            "tokenUrl": "/v1/auth/token"
          }
        }
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}